feat: Add debug check for inverted msb/lsb
In debug mode, check that msb >= lsb in the default `BitRange`/`BitRangeMut` implementations and show a more helpful message than before if this is the case.

This prevents a common error where the user inverts the most and least significant bits, leading to unexpected behavior. The new assertion provides immediate feedback during development.

Thanks to @blaine-dodson for reporting this issue.